Move e2e_args to infra
