+++ tests/code_update.py
@@ -1,6 +1,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the Apache 2.0 License.
-import e2e_args
+import infra.e2e_args
 import infra.ccf
 import infra.path
 import infra.proc
@@ -103,7 +103,7 @@ if __name__ == "__main__":
             "--oesign", help="Path to oesign binary", type=str, required=True
         )
 
-    args = e2e_args.cli_args(add)
+    args = infra.e2e_args.cli_args(add)
     if args.enclave_type != "debug":
         LOG.warning("Skipping code update test with virtual enclave")
         sys.exit()
