+++ tests/connections.py
@@ -4,7 +4,7 @@ import sys
 import os
 import infra.proc
 
-import e2e_args
+import infra.e2e_args
 import getpass
 import os
 import time
@@ -98,5 +98,5 @@ if __name__ == "__main__":
             default="liblogging",
         )
 
-    args = e2e_args.cli_args(add)
+    args = infra.e2e_args.cli_args(add)
     run(args)
