+++ tests/e2e_logging.py
@@ -4,7 +4,7 @@ import infra.ccf
 import infra.notification
 import suite.test_requirements as reqs
 import infra.logging_app as app
-import e2e_args
+import infra.e2e_args
 
 from loguru import logger as LOG
 
@@ -144,7 +144,7 @@ def run(args):
 
 if __name__ == "__main__":
 
-    args = e2e_args.cli_args()
+    args = infra.e2e_args.cli_args()
     if args.js_app_script:
         args.package = "libjsgeneric"
     elif args.app_script:
