+++ tests/e2e_suite.py
@@ -1,7 +1,7 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the Apache 2.0 License.import test_suite
 
-import e2e_args
+import infra.e2e_args
 import infra.ccf
 import suite.test_suite as s
 import suite.test_requirements as reqs
@@ -112,7 +112,7 @@ if __name__ == "__main__":
             "--test-duration", help="Duration of suite of tests (s)", type=int
         )
 
-    args = e2e_args.cli_args(add)
+    args = infra.e2e_args.cli_args(add)
     args.package = args.app_script and "libluageneric" or "liblogging"
 
     run(args)
