+++ tests/loggingclient.py
@@ -1,6 +1,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the Apache 2.0 License.
-import e2e_args
+import infra.e2e_args
 import infra.ccf
 import infra.proc
 
@@ -29,6 +29,6 @@ def run(args):
 
 
 if __name__ == "__main__":
-    args = e2e_args.cli_args()
+    args = infra.e2e_args.cli_args()
     args.package = args.app_script and "libluageneric" or "liblogging"
     run(args)
