+++ tests/node_suspension.py
@@ -11,7 +11,7 @@ import infra.ccf
 import infra.proc
 import infra.notification
 import infra.net
-import e2e_args
+import infra.e2e_args
 from threading import Timer
 import random
 import contextlib
@@ -117,7 +117,7 @@ def run(args):
 
 if __name__ == "__main__":
 
-    args = e2e_args.cli_args()
+    args = infra.e2e_args.cli_args()
     args.package = args.app_script and "libluageneric" or "liblogging"
 
     notify_server_host = "localhost"
