+++ tests/receipts.py
@@ -12,7 +12,7 @@ import infra.proc
 import infra.notification
 import infra.net
 import suite.test_requirements as reqs
-import e2e_args
+import infra.e2e_args
 
 from loguru import logger as LOG
 
@@ -61,7 +61,7 @@ def run(args):
 
 if __name__ == "__main__":
 
-    args = e2e_args.cli_args()
+    args = infra.e2e_args.cli_args()
     args.package = args.app_script or "liblogging"
 
     notify_server_host = "localhost"
