+++ tests/rekey.py
@@ -4,7 +4,7 @@ import infra.ccf
 import infra.jsonrpc
 import infra.notification
 import suite.test_requirements as reqs
-import e2e_args
+import infra.e2e_args
 import time
 
 from loguru import logger as LOG
@@ -57,6 +57,6 @@ def run(args):
 
 if __name__ == "__main__":
 
-    args = e2e_args.cli_args()
+    args = infra.e2e_args.cli_args()
     args.package = args.app_script and "libluageneric" or "liblogging"
     run(args)
