+++ tests/suite/test_requirements.py
@@ -92,10 +92,10 @@ def recover(number_txs=5):
         @functools.wraps(func)
         def wrapper(*args, **kwargs):
             network = args[0]
-            e2e_args = vars(args[1])
+            infra.e2e_args = vars(args[1])
             network.txs.issue(
                 network=network,
-                number_txs=e2e_args.get("msgs_per_recovery") or number_txs,
+                number_txs=infra.e2e_args.get("msgs_per_recovery") or number_txs,
             )
             new_network = func(*args, **kwargs)
             new_network.txs.verify(network=new_network)
